# module5/lab8/calculator/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My Calculator Project",
    "description": "My User Project Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Calculator",
      "description": "API for Calculus in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/calculator/add": {
      "get": {
        "summary": "Adds two numbers",
        "operationId": "addNumbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "description": "The first number",
            "required": true,
            "type": "number"
          },
          {
            "name": "num2",
            "in": "query",
            "description": "The second number",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful addition",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/calculator/subtract": {
      "get": {
        "summary": "Subtracts two numbers",
        "operationId": "subtractNumbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "required": true,
            "type": "number"
          },
          {
            "name": "num2",
            "in": "query",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful subtraction",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/calculator/multiply": {
      "get": {
        "summary": "Multiplies two numbers",
        "operationId": "multiplyNumbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "required": true,
            "type": "number"
          },
          {
            "name": "num2",
            "in": "query",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful multiplication",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/calculator/divide": {
      "get": {
        "summary": "Divides two numbers",
        "operationId": "divideNumbers",
        "parameters": [
          {
            "name": "num1",
            "in": "query",
            "required": true,
            "type": "number"
          },
          {
            "name": "num2",
            "in": "query",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful division",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "number"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or division by zero",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}